Add calculations; separate funcs; small update & fixes
