+++ src/weather_observer.py
@@ -1,5 +1,4 @@
 import argparse
-import codecs
 import logging
 import os
 import sys
@@ -13,6 +12,8 @@ from geopy.geocoders import Nominatim
 from pygismeteo import Gismeteo
 from pytz import timezone
 from timezonefinder import TimezoneFinder
+
+import calculations
 
 # Logging
 logging.basicConfig(
@@ -31,14 +32,14 @@ REPORT_NAME = "weather_report_"
 REPORT_FORMAT = ".md"
 
 # Time when program execution started
-START_TIME = time.time()
+start_time = time.time()
 
 WEATHER_API = "https://api.weatherbit.io/v2.0/"
 # Using in get_current_city func to retrieve current city name
 IP_SITE = "http://ipinfo.io/"
 OPEN_ELEVATION_API = "https://api.open-elevation.com/api/v1/lookup?locations="
 
-REPORT_TIME = datetime.now().strftime("%d.%m.%Y_%H.%M.%S")
+report_time = datetime.now().strftime("%d.%m.%Y_%H.%M.%S")
 
 # Conversion for pressure
 KPA = 0.1  # Kilo Pascal
@@ -164,7 +165,7 @@ def get_time_by_timezone(timezone_name: str) -> str:
     return now_timezone.strftime(date_time_format)
 
 
-def request_weather_info(country_code: str, city_name: str) -> Dict:
+def request_weather_info(country_code: str, city_name: str) -> Dict[str, any]:
     """
     Send GET request to weatherbit resource fetching info about weather about transferred countries & cities
     :param country_code:
@@ -177,111 +178,22 @@ def request_weather_info(country_code: str, city_name: str) -> Dict:
         )
         return r.json()["data"][0]
     except requests.exceptions.RequestException as req_ex:
-        logging.error(req_ex)
+        logging.error(f"Err while request weather info from API - {req_ex}")
+        return None
     except BaseException as base_err:
-        logging.error(base_err)
+        logging.error(f"Base err while request weather info from API - {base_err}")
+        return None
 
 
-def calculate_uv_level(uv_value: float) -> str:
-    """
-    Passed UV value as float and return string value on the scale
-    :param uv_value:
-    :return:
-    """
-    if 0.0 <= uv_value <= 2.9:
-        return "green"
-    elif 3.0 <= uv_value <= 5.9:
-        return "yellow"
-    elif 6.0 <= uv_value <= 7.9:
-        return "orange"
-    elif 8.0 <= uv_value <= 10.9:
-        return "red"
-    elif 11.0 <= uv_value:
-        return "purple"
-
-
-def calculate_aqi_level(aqi_value: int) -> str:
-    """
-    Passed Air Quality Index value as integer and return string value on the scale
-    :param aqi_value:
-    :return:
-    """
-    if 0 <= aqi_value <= 33:
-        return "very good"
-    elif 34 <= aqi_value <= 66:
-        return "good"
-    elif 67 <= aqi_value <= 99:
-        return "fair"
-    elif 100 <= aqi_value <= 149:
-        return "poor"
-    elif 150 <= aqi_value <= 200:
-        return "very poor"
-    elif 200 <= aqi_value:
-        return "hazardous"
-
-
-def calculate_kp_level(kp_value: int) -> str:
-    """
-    Passed Kp value as int and return string value on the scale
-    The Kp-index describes the disturbance of the Earth’s magnetic field caused by the solar wind
-    :param kp_value:
-    :return:
-    """
-    if 0 <= kp_value < 3:
-        return "quiet"
-    elif kp_value == 3:
-        return "unsettled"
-    elif kp_value == 4:
-        return "active"
-    elif kp_value == 5:
-        return "minor storm"
-    elif kp_value == 6:
-        return "moderate storm"
-    elif kp_value == 7:
-        return "strong storm"
-    elif kp_value == 8:
-        return "severe storm"
-    elif kp_value >= 9:
-        return "intense storm"
-
-
-def celsius_to_fahrenheit(celsius: float) -> float:
-    """
-    Convert celsius to fahrenheit
-    :param celsius:
-    :return:
-    """
-    return (celsius * 9 / 5) + 32
-
-
-def celsius_to_kelvin(celsius: float) -> float:
-    """
-    Convert celsius to kelvin
-    :param celsius:
-    :return:
-    """
-    return celsius + 273.15
-
-
-def prepare_weather_info(
-    country_name: str,
+def prepare_weather_data(
     country_code: str,
     city_name: str,
-    timezone_by_city: str,
-    elevation: int,
-    water_temp: float,
-    geomagnetic_field: int,
 ):
     """
     Prepare weather information to better writing into report file
     After result is ready, pass it to the report_weather_info func
-    :param country_name:
     :param country_code:
     :param city_name:
-    :param timezone_by_city:
-    :param elevation:
-    :param water_temp:
-    :param geomagnetic_field:
     :return:
     """
 
@@ -289,16 +201,7 @@ def prepare_weather_info(
     for v in VALUES_TO_DELETE:
         del result[v]
 
-    report_weather_info(
-        REPORT_TIME,
-        result,
-        city_name,
-        timezone_by_city,
-        country_name,
-        elevation,
-        water_temp,
-        geomagnetic_field,
-    )
+    return result
 
 
 def report_to_console(
@@ -329,7 +232,7 @@ def report_to_console(
     print(f"Part of a day: {weather_data['pod']}")
     print(f"Elevation above sea level: {elevation} m")
     print(
-        f"Geomagnetic field: {geomagnetic_field} - {calculate_kp_level(geomagnetic_field).capitalize()}"
+        f"Geomagnetic field: {geomagnetic_field} - {calculations.calculate_kp_level(geomagnetic_field).capitalize()}"
     )
     print()
     print(
@@ -352,27 +255,27 @@ def report_to_console(
     print()
     print(
         f"UV (UltraViolet): {weather_data['uv']} - "
-        f"{calculate_uv_level(round(weather_data['uv'], 1)).capitalize()}"
+        f"{calculations.calculate_uv_level(round(weather_data['uv'], 1)).capitalize()}"
     )
     print(
         f"AQI (Air Quality Index): {weather_data['aqi']} - "
-        f"{calculate_aqi_level(weather_data['aqi']).capitalize()}"
+        f"{calculations.calculate_aqi_level(weather_data['aqi']).capitalize()}"
     )
     print()
     print(
         f"Temperature: {weather_data['temp']} C "
-        f"| {round(celsius_to_fahrenheit(weather_data['temp']), 1)} F "
-        f"| {round(celsius_to_kelvin(weather_data['temp']), 1)} K"
+        f"| {round(calculations.celsius_to_fahrenheit(weather_data['temp']), 1)} F "
+        f"| {round(calculations.celsius_to_kelvin(weather_data['temp']), 1)} K"
     )
     print(
         f"Apparent temperature: {weather_data['app_temp']} C "
-        f"| {round(celsius_to_fahrenheit(weather_data['app_temp']), 1)} F "
-        f"| {round(celsius_to_kelvin(weather_data['app_temp']), 1)} K "
+        f"| {round(calculations.celsius_to_fahrenheit(weather_data['app_temp']), 1)} F "
+        f"| {round(calculations.celsius_to_kelvin(weather_data['app_temp']), 1)} K "
     )
     print(
         f"Water temperature in location: {water_temp} C "
-        f"| {round(celsius_to_fahrenheit(water_temp), 1)} F "
-        f"| {round(celsius_to_kelvin(water_temp), 1)} K  "
+        f"| {round(calculations.celsius_to_fahrenheit(water_temp), 1)} F "
+        f"| {round(calculations.celsius_to_kelvin(water_temp), 1)} K  "
     )
     input("Enter any key to escape...")
 
@@ -421,7 +324,7 @@ def report_to_telegram(
                     f"Elevation under sea level: {elevation} m"
                     f"\n"
                     f"Geomagnetic field: {geomagnetic_field} - "
-                    f"{calculate_kp_level(geomagnetic_field).capitalize()}"
+                    f"{calculations.calculate_kp_level(geomagnetic_field).capitalize()}"
                     f"\n"
                     f"\n"
                     f"Pressure: {round(weather_data['pres'], 2)} mb "
@@ -447,23 +350,23 @@ def report_to_telegram(
                     f"\n"
                     f"\n"
                     f"UV (UltraViolet): {weather_data['uv']} - "
-                    f"{calculate_uv_level(round(weather_data['uv'], 1)).capitalize()}"
+                    f"{calculations.calculate_uv_level(round(weather_data['uv'], 1)).capitalize()}"
                     f"\n"
                     f"AQI (Air Quality Index): {weather_data['aqi']} - "
-                    f"{calculate_aqi_level(weather_data['aqi']).capitalize()}"
+                    f"{calculations.calculate_aqi_level(weather_data['aqi']).capitalize()}"
                     f"\n"
                     f"\n"
                     f"Temperature: {weather_data['temp']} C "
-                    f"| {round(celsius_to_fahrenheit(weather_data['temp']), 1)} F "
-                    f"| {round(celsius_to_kelvin(weather_data['temp']), 1)} K"
+                    f"| {round(calculations.celsius_to_fahrenheit(weather_data['temp']), 1)} F "
+                    f"| {round(calculations.celsius_to_kelvin(weather_data['temp']), 1)} K"
                     f"\n"
                     f"Apparent temperature: {weather_data['app_temp']} C "
-                    f"| {round(celsius_to_fahrenheit(weather_data['app_temp']), 1)} F "
-                    f"| {round(celsius_to_kelvin(weather_data['app_temp']), 1)} K "
+                    f"| {round(calculations.celsius_to_fahrenheit(weather_data['app_temp']), 1)} F "
+                    f"| {round(calculations.celsius_to_kelvin(weather_data['app_temp']), 1)} K "
                     f"\n"
                     f"Water temperature in location: {water_temp} C "
-                    f"| {round(celsius_to_fahrenheit(water_temp), 1)} F "
-                    f"| {round(celsius_to_kelvin(water_temp),1)} K  "
+                    f"| {round(calculations.celsius_to_fahrenheit(water_temp), 1)} F "
+                    f"| {round(calculations.celsius_to_kelvin(water_temp), 1)} K  "
                     f"\n",
                 },
             )
@@ -505,8 +408,8 @@ def report_to_file(
     :param geomagnetic_field:
     :return:
     """
-    with codecs.open(
-        f"{REPORT_NAME}{report_time}{REPORT_FORMAT}", "a", "utf-8"
+    with open(
+        f"{REPORT_NAME}{report_time}{REPORT_FORMAT}", "a", encoding="utf-8"
     ) as report:
         if namespace.verbosity:
             print(f"Gathering info about {city_name.capitalize()} in {country_name}...")
@@ -519,7 +422,7 @@ def report_to_file(
         )
         report.write(f"**Elevation under sea level:** {elevation} m  \n")
         report.write(
-            f"Geomagnetic field: {geomagnetic_field} - {calculate_kp_level(geomagnetic_field).capitalize()}  \n"
+            f"Geomagnetic field: {geomagnetic_field} - {calculations.calculate_kp_level(geomagnetic_field).capitalize()}  \n"
         )
         report.write(
             f"Country: {country_name} | City name: {city_name.capitalize()}  \n"
@@ -532,7 +435,7 @@ def report_to_file(
         report.write(f"Part of a day: {weather_data['pod']}  \n")
         report.write(f"Elevation above sea level: {elevation} m  \n")
         report.write(
-            f"Geomagnetic field: {geomagnetic_field} - {calculate_kp_level(geomagnetic_field).capitalize()}  \n"
+            f"Geomagnetic field: {geomagnetic_field} - {calculations.calculate_kp_level(geomagnetic_field).capitalize()}  \n"
         )
         report.write("\n")
         report.write(
@@ -555,30 +458,30 @@ def report_to_file(
         report.write("\n")
         report.write(
             f"UV (UltraViolet): {weather_data['uv']} - "
-            f"{calculate_uv_level(round(weather_data['uv'], 1)).capitalize()}  \n"
+            f"{calculations.calculate_uv_level(round(weather_data['uv'], 1)).capitalize()}  \n"
         )
         report.write(
             f"AQI (Air Quality Index): {weather_data['aqi']} - "
-            f"{calculate_aqi_level(weather_data['aqi']).capitalize()}  \n"
+            f"{calculations.calculate_aqi_level(weather_data['aqi']).capitalize()}  \n"
         )
         report.write("\n")
         report.write(
             f"**Temperature**: {weather_data['temp']} C "
-            f"| {round(celsius_to_fahrenheit(weather_data['temp']), 1)} F "
-            f"| {round(celsius_to_kelvin(weather_data['temp']), 1)} K  \n"
+            f"| {round(calculations.celsius_to_fahrenheit(weather_data['temp']), 1)} F "
+            f"| {round(calculations.celsius_to_kelvin(weather_data['temp']), 1)} K  \n"
         )
         report.write(
             f"**Apparent temperature**: {weather_data['app_temp']} C "
-            f"| {round(celsius_to_fahrenheit(weather_data['app_temp']), 1)} F "
-            f"| {round(celsius_to_kelvin(weather_data['app_temp']), 1)} K  \n"
+            f"| {round(calculations.celsius_to_fahrenheit(weather_data['app_temp']), 1)} F "
+            f"| {round(calculations.celsius_to_kelvin(weather_data['app_temp']), 1)} K  \n"
         )
         report.write(
             f"**Water temperature in location**: {water_temp} C "
-            f"| {round(celsius_to_fahrenheit(water_temp), 1)} F "
-            f"| {round(celsius_to_kelvin(water_temp), 1)} K  \n"
+            f"| {round(calculations.celsius_to_fahrenheit(water_temp), 1)} F "
+            f"| {round(calculations.celsius_to_kelvin(water_temp), 1)} K  \n"
         )
         if namespace.verbosity:
-            print("--- %s seconds ---" % (time.time() - START_TIME))
+            print("--- %s seconds ---" % (time.time() - start_time))
         report.write("\n")
 
 
@@ -755,30 +658,34 @@ def prepare_target_location_info(city_name: str):
                 "22:00:00",
             ]:
                 logging.info("It is time to report !")
-                prepare_weather_info(
-                    country_name,
-                    country_code,
-                    city_name,
-                    timezone_by_city,
-                    get_elevation_by_ll(latitude=latitude, longitude=longitude),
-                    get_water_temp_by_ll(
+                report_weather_info(
+                    report_time=report_time,
+                    weather_data=prepare_weather_data(country_code, city_name),
+                    city_name=city_name,
+                    timezone_by_city=timezone_by_city,
+                    country_name=country_name,
+                    elevation=get_elevation_by_ll(
+                        latitude=latitude, longitude=longitude
+                    ),
+                    water_temp=get_water_temp_by_ll(
                         latitude=location.latitude, longitude=location.longitude
                     ),
-                    get_geomagnetic_field_by_ll(
+                    geomagnetic_field=get_geomagnetic_field_by_ll(
                         latitude=location.latitude, longitude=location.longitude
                     ),
                 )
     else:
-        prepare_weather_info(
-            country_name,
-            country_code,
-            city_name,
-            timezone_by_city,
-            get_elevation_by_ll(latitude=latitude, longitude=longitude),
-            get_water_temp_by_ll(
+        report_weather_info(
+            report_time=report_time,
+            weather_data=prepare_weather_data(country_code, city_name),
+            city_name=city_name,
+            timezone_by_city=timezone_by_city,
+            country_name=country_name,
+            elevation=get_elevation_by_ll(latitude=latitude, longitude=longitude),
+            water_temp=get_water_temp_by_ll(
                 latitude=location.latitude, longitude=location.longitude
             ),
-            get_geomagnetic_field_by_ll(
+            geomagnetic_field=get_geomagnetic_field_by_ll(
                 latitude=location.latitude, longitude=location.longitude
             ),
         )
